+++ src/main/java/com/contentful/java/cma/ModuleAssets.java
@@ -97,15 +97,15 @@ public final class ModuleAssets extends AbsModule<ServiceAssets> {
    *
    * @param spaceId Space ID
    * @param assetId Asset ID
-   * @return A string representing the result of the delete operation
+   * @return An integer representing the result of the delete operation
    * @throws IllegalArgumentException if spaceId is null.
    * @throws IllegalArgumentException if assetId is null.
    */
-  public String delete(String spaceId, String assetId) {
+  public Integer delete(String spaceId, String assetId) {
     assertNotNull(spaceId, "spaceId");
     assertNotNull(assetId, "assetId");
 
-    return service.delete(spaceId, assetId).blockingFirst();
+    return service.delete(spaceId, assetId).blockingFirst().code();
   }
 
   /**
@@ -157,18 +157,18 @@ public final class ModuleAssets extends AbsModule<ServiceAssets> {
    *
    * @param asset  Asset
    * @param locale Locale
-   * @return String representing the success (203) of processing
+   * @return integer representing the success (204 = no content) of processing
    * @throws IllegalArgumentException if asset is null.
    * @throws IllegalArgumentException if asset has no id.
    * @throws IllegalArgumentException if asset has no space.
    * @throws IllegalArgumentException if locale is null.
    */
-  public String process(CMAAsset asset, String locale) {
+  public Integer process(CMAAsset asset, String locale) {
     assertNotNull(asset, "asset");
     final String assetId = getResourceIdOrThrow(asset, "asset");
     final String spaceId = getSpaceIdOrThrow(asset, "asset");
 
-    return service.process(spaceId, assetId, locale).blockingFirst();
+    return service.process(spaceId, assetId, locale).blockingFirst().code();
   }
 
   /**
@@ -311,10 +311,11 @@ public final class ModuleAssets extends AbsModule<ServiceAssets> {
      * @throws IllegalArgumentException if spaceId is null.
      * @throws IllegalArgumentException if assetId is null.
      */
-    public CMACallback<String> delete(final String spaceId, final String assetId,
-                                      CMACallback<String> callback) {
-      return defer(new DefFunc<String>() {
-        @Override String method() {
+    public CMACallback<Integer> delete(final String spaceId,
+                                       final String assetId,
+                                       CMACallback<Integer> callback) {
+      return defer(new DefFunc<Integer>() {
+        @Override Integer method() {
           return ModuleAssets.this.delete(spaceId, assetId);
         }
       }, callback);
@@ -383,10 +384,10 @@ public final class ModuleAssets extends AbsModule<ServiceAssets> {
      * @throws IllegalArgumentException if asset has no space.
      * @throws IllegalArgumentException if locale is null.
      */
-    public CMACallback<String> process(final CMAAsset asset, final String locale,
-                                       CMACallback<String> callback) {
-      return defer(new DefFunc<String>() {
-        @Override String method() {
+    public CMACallback<Integer> process(final CMAAsset asset, final String locale,
+                                        CMACallback<Integer> callback) {
+      return defer(new DefFunc<Integer>() {
+        @Override Integer method() {
           return ModuleAssets.this.process(asset, locale);
         }
       }, callback);